test: consolidate battleCLI a11y smoke
